Refactor: Remove BookshelfMatchModalHost and integrate its functionality into MatchModal
